requirements.txt
webdavclient3==3.14.6
python-dotenv==1.0.1
typer==0.12.5